MAINT: address comments on PR

+++ src/cogent3/core/new_alphabet.py
@@ -326,7 +326,7 @@ class CharAlphabet(tuple, AlphabetABC, MonomerAlphabetABC):
 
     @to_indices.register
     def _(self, seq: numpy.ndarray) -> numpy.ndarray:
-        return seq.astype(numpy.uint8)
+        return seq.astype(self.dtype)
 
     @functools.singledispatchmethod
     def from_indices(self, seq: numpy.ndarray) -> str:
